Remove types file
