+++ src/decide.ts
@@ -3,8 +3,14 @@ import * as cache from "./cache";
 import { EP_REGEX, MOVIE_REGEX, SEASON_REGEX } from "./constants";
 import * as logger from "./logger";
 import { partial } from "./utils";
-import { JackettResult, ResultAssessment } from "./types";
 import { Metafile } from "parse-torrent";
+import { JackettResult } from "./jackett";
+
+export interface ResultAssessment {
+	tracker: string;
+	tag: string;
+	info: Metafile;
+}
 
 function compareFileTrees(a, b) {
 	const cmp = (elOfA, elOfB) => {
